remove all tooltips for iPad and iPhon #465, #914 remove brickname from configuration (not valid anymore), show annotated blocks on error, revise toolbox'
