+++ OpenRobertaParent/OpenRobertaServer/staticResources/index.html
@@ -730,16 +730,15 @@ limitations under the License.
             </div>
         </div>
     </div>
-    <!--  Popup to show relations / share programs -->
+    <!--  Popup to show connections triggered from a webview-->
     <div id="show-available-connections" class="modal fade">
         <div class="modal-dialog">
             <div class="modal-content">
                 <div class="modal-header">
                     <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
-                    <h3 id="#headConnections" class="modal-title">Available robots</h3>
+                    <h3 id="#headConnections" class="modal-title" lkey="Blockly.Msg.POPUP_SCANNED_ROBOTS">Gefundene Roboter</h3>
                 </div>
-                <!-- <form id="share-program-form"> -->
-                <div class="modal-body">
+                <div class="modal-body" style="height: 158px;">
                     <table id="connectionsTable">
                     </table>
                 </div>
