+++ OpenRobertaParent/OpenRobertaServer/staticResources/js/app/roberta/controller/webview.controller.js
@@ -59,7 +59,6 @@ define([ 'exports', 'guiState.controller', 'wedo.model', 'interpreter.interprete
                             var dom = Blockly.Xml.workspaceToDom(bricklyWorkspace);
                             var xml = Blockly.Xml.domToText(dom);
                             GUISTATE_C.setConfigurationXML(xml);
-                            bricklyWorkspace.setVisible(false);
                             break;
                         }
                     }
